fix: 네이밍 변경에 따른 책임 분리
- 책임분리를 위해 나이, 성별 수정 메서드를 user service로 이동
